(#19663) Remove --include-dependencies flag in gem package provider
Without this patch applied the gem package provider will always pass the
`--include-dependencies` flag to the gem command.  This flag has been
deprecated in Rubygems for quite some time.  The behavior of including
dependencies when installing a gem has been the default for some time,
so this patch does not change behavior for any recent version of
Rubygems since 0.9.5.

Here is the comment from the pull request:

This is a good explanation about the option: rubygems/rubygems#385

It was already deprecated since rubygems 0.9.5 and was the default
option since then.

Unless someone is using a version of rubygems from mid 2007, this option
is never being used
https://github.com/rubygems/rubygems/blob/a00fc76fb6b959385331715c500af1eb25590405/History.txt#L1451